+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/__init__.py
@@ -60,6 +60,11 @@ from ._models import (
     ExtractSummaryAction,
     ExtractSummaryResult,
     SummarySentence,
+    AbstractSummaryResult,
+    AbstractiveSummary,
+    SummaryContext,
+    PhraseControl,
+    AbstractSummaryAction,
 )
 from ._generated.models import (
     HealthcareDocumentType,
@@ -92,6 +97,7 @@ from ._generated.models import (
     TemperatureUnit,
     VolumeUnit,
     WeightUnit,
+    PhraseControlStrategy,
 )
 from ._lro import AnalyzeHealthcareEntitiesLROPoller, AnalyzeActionsLROPoller, TextAnalysisLROPoller
 
@@ -184,6 +190,12 @@ __all__ = [
     "TemperatureUnit",
     "VolumeUnit",
     "WeightUnit",
+    "AbstractSummaryResult",
+    "AbstractiveSummary",
+    "SummaryContext",
+    "PhraseControlStrategy",
+    "PhraseControl",
+    "AbstractSummaryAction"
 ]
 
 __version__ = VERSION