+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_response_handlers.py
@@ -37,6 +37,7 @@ from ._models import (
     ClassifyDocumentResult,
     ActionPointerKind,
     ExtractSummaryResult,
+    AbstractSummaryResult,
 )
 
 
@@ -128,6 +129,15 @@ def prepare_result(func):
         return wrapper(*args, ordering_function=order_results)
 
     return choose_wrapper
+
+
+@prepare_result
+def abstract_summary_result(
+    summary, results, *args, **kwargs
+):  # pylint: disable=unused-argument
+    return AbstractSummaryResult._from_generated(  # pylint: disable=protected-access
+        summary
+    )
 
 
 @prepare_result
@@ -324,6 +334,8 @@ def _get_deserialization_callback_from_task_type(task_type):  # pylint: disable=
         return healthcare_result
     if task_type == _AnalyzeActionsType.EXTRACT_SUMMARY:
         return summary_result
+    if task_type == _AnalyzeActionsType.ABSTRACT_SUMMARY:
+        return abstract_summary_result
     return key_phrases_result