+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_text_analytics_client.py
@@ -67,6 +67,8 @@ from ._models import (
     _AnalyzeActionsType,
     ExtractSummaryAction,
     ExtractSummaryResult,
+    AbstractSummaryAction,
+    AbstractSummaryResult,
 )
 from ._check import is_language_api, string_index_type_compatibility
 
@@ -84,6 +86,7 @@ AnalyzeActionsResponse = TextAnalysisLROPoller[
                 ClassifyDocumentResult,
                 AnalyzeHealthcareEntitiesResult,
                 ExtractSummaryResult,
+                AbstractSummaryResult,
                 DocumentError,
             ]
         ]
@@ -1079,6 +1082,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                 MultiLabelClassifyAction,
                 AnalyzeHealthcareEntitiesAction,
                 ExtractSummaryAction,
+                AbstractSummaryAction,
             ]
         ],
         **kwargs: Any,
@@ -1095,6 +1099,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                     ClassifyDocumentResult,
                     AnalyzeHealthcareEntitiesResult,
                     ExtractSummaryResult,
+                    AbstractSummaryResult,
                     DocumentError,
                 ]
             ]
@@ -1123,7 +1128,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             list[RecognizeEntitiesAction or RecognizePiiEntitiesAction or ExtractKeyPhrasesAction or
             RecognizeLinkedEntitiesAction or AnalyzeSentimentAction or
             RecognizeCustomEntitiesAction or SingleLabelClassifyAction or
-            MultiLabelClassifyAction or AnalyzeHealthcareEntitiesAction or ExtractSummaryAction]
+            MultiLabelClassifyAction or AnalyzeHealthcareEntitiesAction or ExtractSummaryAction
+            or AbstractSummaryAction]
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
@@ -1152,7 +1158,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             ~azure.ai.textanalytics.TextAnalysisLROPoller[~azure.core.paging.ItemPaged[
             list[RecognizeEntitiesResult or RecognizeLinkedEntitiesResult or RecognizePiiEntitiesResult or
             ExtractKeyPhrasesResult or AnalyzeSentimentResult or RecognizeCustomEntitiesResult
-            or ClassifyDocumentResult or AnalyzeHealthcareEntitiesResult or ExtractSummaryResult or DocumentError]]]
+            or ClassifyDocumentResult or AnalyzeHealthcareEntitiesResult or ExtractSummaryResult
+            or AbstractSummaryResult or DocumentError]]]
         :raises ~azure.core.exceptions.HttpResponseError or TypeError or ValueError:
 
         .. versionadded:: v3.1
@@ -1163,8 +1170,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             corresponding *RecognizeCustomEntitiesResult*, *ClassifyDocumentResult*,
             and *AnalyzeHealthcareEntitiesResult* result objects
         .. versionadded:: 2022-10-01-preview
-            The *ExtractSummaryAction* input option and the corresponding *ExtractSummaryResult*
-            result object.
+            The *ExtractSummaryAction* and *AbstractSummaryAction* input options and the corresponding
+            *ExtractSummaryResult* and *AbstractSummaryResult* result objects.
 
         .. admonition:: Example: