+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_validate.py
@@ -30,6 +30,7 @@ def check_for_unsupported_actions_types(*args, **kwargs):
         "2022-10-01-preview":
         [
             "ExtractSummaryAction",
+            "AbstractSummaryAction",
         ],
         "2022-05-01":
         [