+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze.py
@@ -41,7 +41,9 @@ from azure.ai.textanalytics import (
     RecognizeCustomEntitiesResult,
     AnalyzeHealthcareEntitiesAction,
     ExtractSummaryAction,
-    ExtractSummaryResult
+    ExtractSummaryResult,
+    AbstractSummaryAction,
+    PhraseControl,
 )
 
 # pre-apply the client_cls positional argument so it needn't be explicitly passed below
@@ -1940,3 +1942,81 @@ class TestAnalyze(TextAnalyticsTest):
                         assert res.value == 10
                     if res.resolution_kind == "AgeResolution":
                         assert res.value == 1
+
+    @TextAnalyticsPreparer()
+    @TextAnalyticsClientPreparer()
+    @recorded_by_proxy
+    def test_passing_dict_abstract_summary_action(self, client):
+        docs = [{"id": "1", "language": "en", "text":
+            "The government of British Prime Minster Theresa May has been plunged into turmoil with the resignation"
+            " of two senior Cabinet ministers in a deep split over her Brexit strategy. The Foreign Secretary Boris "
+            "Johnson, quit on Monday, hours after the resignation late on Sunday night of the minister in charge of "
+            "Brexit negotiations, David Davis. Their decision to leave the government came three days after May "
+            "appeared to have agreed a deal with her fractured Cabinet on the UK's post Brexit relationship with "
+            "the EU. That plan is now in tatters and her political future appears uncertain. May appeared in Parliament"
+            " on Monday afternoon to defend her plan, minutes after Downing Street confirmed the departure of Johnson. "
+            "May acknowledged the splits in her statement to MPs, saying of the ministers who quit: We do not agree "
+            "about the best way of delivering our shared commitment to honoring the result of the referendum. The "
+            "Prime Minister's latest political drama began late on Sunday night when Davis quit, declaring he could "
+            "not support May's Brexit plan. He said it involved too close a relationship with the EU and gave only "
+            "an illusion of control being returned to the UK after it left the EU. It seems to me we're giving too "
+            "much away, too easily, and that's a dangerous strategy at this time, Davis said in a BBC radio "
+            "interview Monday morning. Johnson's resignation came Monday afternoon local time, just before the "
+            "Prime Minister was due to make a scheduled statement in Parliament. This afternoon, the Prime Minister "
+            "accepted the resignation of Boris Johnson as Foreign Secretary, a statement from Downing Street said."}]
+
+        response = client.begin_analyze_actions(
+            docs,
+            actions=[AbstractSummaryAction()],
+            show_stats=True,
+            polling_interval=self._interval(),
+        ).result()
+
+        document_results = list(response)
+
+        for document_result in document_results:
+            for result in document_result:
+                assert result.statistics is not None
+                assert result.id is not None
+                for summary in result.summaries:
+                    for context in summary.contexts:
+                        assert context.offset is not None
+                        assert context.length is not None
+                    assert summary.text
+
+    @pytest.mark.skip("https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15919116")
+    @TextAnalyticsPreparer()
+    @TextAnalyticsClientPreparer()
+    @recorded_by_proxy
+    def test_abstract_summary_action_with_options(self, client):
+        docs = [{"id": "1", "language": "en", "text": "At Microsoft, we have been on a quest to advance AI beyond existing techniques, by taking a more holistic, human-centric approach to learning and understanding. As Chief Technology Officer of Azure AI Cognitive Services, I have been working with a team of amazing scientists and engineers to turn this quest into a reality. In my role, I enjoy a unique perspective in viewing the relationship among three attributes of human cognition: monolingual text (X), audio or visual sensory signals, (Y) and multilingual (Z). At the intersection of all three, there’s magic—what we call XYZ-code as illustrated in Figure 1—a joint representation to create more powerful AI that can speak, hear, see, and understand humans better. We believe XYZ-code will enable us to fulfill our long-term vision: cross-domain transfer learning, spanning modalities and languages. The goal is to have pre-trained models that can jointly learn representations to support a broad range of downstream AI tasks, much in the way humans do today. Over the past five years, we have achieved human performance on benchmarks in conversational speech recognition, machine translation, conversational question answering, machine reading comprehension, and image captioning. These five breakthroughs provided us with strong signals toward our more ambitious aspiration to produce a leap in AI capabilities, achieving multi-sensory and multilingual learning that is closer in line with how humans learn and understand. I believe the joint XYZ-code is a foundational component of this aspiration, if grounded with external knowledge sources in the downstream AI tasks."}]
+
+        response = client.begin_analyze_actions(
+            docs,
+            actions=[
+                AbstractSummaryAction(
+                    sentence_count=4,
+                    phrase_controls=[
+                        PhraseControl(
+                            target_phrase="Microsoft",
+                            strategy="encourage"
+                        )
+                    ]
+                )
+            ],
+            show_stats=True,
+            polling_interval=self._interval(),
+        ).result()
+
+        document_results = list(response)
+
+        for document_result in document_results:
+            for result in document_result:
+                assert result.statistics is not None
+                assert result.id is not None
+                for summary in result.summaries:
+                    for context in summary.contexts:
+                        assert context.offset is not None
+                        assert context.length is not None
+                    assert summary.text
+